@@ -23,18 +23,17 @@
     ]
   },
   "require": {
-    "php": "^7.2",
+    "php": "^7.4",
     "ext-pcntl": "*",
     "ext-json": "*",
     "yiisoft/yii2": "^2.0",
     "yiisoft/yii2-swiftmailer": "^2.1",
-    "tebazil/yii2-db-seeder": "^0.0.0",
     "yiisoft/yii2-authclient": "^2.2.4",
     "vlucas/phpdotenv": "^3.3",
-    "manchenkov/yii2-helpers": "^1.0",
-    "manchenkov/php-helpers": "^1.0",
-    "manchenkov/yii2-recaptcha3": "^1.0",
-    "manchenkov/yii2-components": "^1.0.14"
+    "manchenkov/yii2-helpers": "^1.0.9",
+    "manchenkov/php-helpers": "^1.0.9",
+    "manchenkov/yii2-recaptcha3": "^1.0.2",
+    "manchenkov/yii2-components": "^2.0.1"
   },
   "require-dev": {
     "deployer/deployer": "^6.4",
@@ -43,6 +42,7 @@
     "guzzlehttp/guzzle": "^6.3",
     "symfony/event-dispatcher": "<=4.2.8",
     "symfony/browser-kit": ">=2.7 <=4.2.4",
-    "manchenkov/yii2-maker": "^1.0"
+    "tebazil/yii2-db-seeder": "^0.0.0",
+    "manchenkov/yii2-maker": "^1.0.4"
   }
 }
